+++ metric/duration_point.go
@@ -7,22 +7,22 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/cloudwatch"
 )
 
-type DurationPointMarker interface {
-	Point
-	Mark()
-}
-
 // This will report duration in microseconds to cloudwatch
 // The duration starts when this is created and ends when converted to MetricDatum via ToAWS
-func DurationPoint(name string) DurationPointMarker {
+func DurationPoint(name string, dimensions ...Dimension) PointMarker {
 	return &durationPoint{
-		name:  name,
+		basePoint: basePoint{
+			name:       name,
+			unit:       cloudwatch.StandardUnitMicroseconds,
+			dimensions: dimensions,
+			timestamp:  time.Now(),
+		},
 		start: time.Now(),
 	}
 }
 
 type durationPoint struct {
-	name  string
+	basePoint
 	start time.Time
 	dur   *time.Duration
 }
@@ -36,11 +36,7 @@ func (p *durationPoint) Mark() {
 
 func (p *durationPoint) ToAWS() cloudwatch.MetricDatum {
 	p.Mark()
-	return cloudwatch.MetricDatum{
-		Unit:              cloudwatch.StandardUnitMicroseconds,
-		MetricName:        aws.String(p.name),
-		Value:             aws.Float64(float64(p.dur.Nanoseconds() / 1000)),
-		StorageResolution: aws.Int64(1),
-		Timestamp:         aws.Time(time.Now()),
-	}
+	md := p.basePoint.ToAWS()
+	md.Value = aws.Float64(float64(p.dur.Nanoseconds() / 1000))
+	return md
 }